+++ client/styles/account/register.ts
